Simplified front-end cookie usage. userCookie will now always be sent with requests as a header
