+++ harmony-fe/src/contexts/user.ts
@@ -2,4 +2,3 @@ import React from 'react'
 import { User } from '../models/user'
 
 export const UserContext = React.createContext<User | undefined>(undefined)
-export const UserCookieContext = React.createContext<string>('')
