+++ harmony-fe/src/pages/_app.tsx
@@ -2,7 +2,7 @@ import '../styles/global.css'
 import type { ReactElement, ReactNode } from 'react'
 import type { Metadata, NextPage } from 'next'
 import type { AppProps } from 'next/app'
-import { UserContext, UserCookieContext } from '../contexts/user'
+import { UserContext } from '../contexts/user'
 import { MobileContext } from '../contexts/mobile'
 import useHttpRequest, { HttpMethod } from '../hooks/httpRequest'
 import { useEffect, useState } from 'react'
@@ -52,34 +52,16 @@ type AppPropsWithLayout = AppProps & {
 
 export default function App({ Component, pageProps }: AppPropsWithLayout) {
   const [currentUser, setCurrentUser] = useState<User | undefined>(undefined)
-  const [userCookie, setUserCookie] = useState<string>('')
   const [mobile, setMobile] = useState<boolean>(false)
-  const cookieInfo = { userCookie: '' }
-  function getCookie(cookieName: string): string {
-    const name = cookieName + '='
-    const decodedCookie = decodeURIComponent(document.cookie)
-    const cookieList = decodedCookie.split(';')
-    let foundCookie = ''
-    cookieList.forEach(val => {
-      if (val.indexOf(name) === 0) {
-        foundCookie = val.substring(name.length)
-      }
-    })
-    return foundCookie
-  }
+
 
   const [sendHttpRequest, response, error, loading] = useHttpRequest({
     url: '/user/getUser',
     method: HttpMethod.GET,
-    headers: cookieInfo,
     body: '',
   })
   useEffect(() => {
-    cookieInfo.userCookie = getCookie('userCookie')
-    setUserCookie(cookieInfo.userCookie as string)
-    if (cookieInfo.userCookie !== '') {
       sendHttpRequest()
-    }
   }, [])
   useEffect(() => {
     const userAgent = window.navigator.userAgent
@@ -99,16 +81,14 @@ export default function App({ Component, pageProps }: AppPropsWithLayout) {
   const getLayout = Component.getLayout ?? (page => page)
 
   return getLayout(
-    <UserCookieContext.Provider value={userCookie}>
-      <UserContext.Provider value={currentUser}>
-        <MobileContext.Provider value={mobile}>
-          <ThemeProvider theme={globalTheme}>
-            <CssBaseline />
-            <HarmonyAppBar />
-            <Component {...pageProps} />
-          </ThemeProvider>
-        </MobileContext.Provider>
-      </UserContext.Provider>
-    </UserCookieContext.Provider>
+    <UserContext.Provider value={currentUser}>
+      <MobileContext.Provider value={mobile}>
+        <ThemeProvider theme={globalTheme}>
+          <CssBaseline />
+          <HarmonyAppBar />
+          <Component {...pageProps} />
+        </ThemeProvider>
+      </MobileContext.Provider>
+    </UserContext.Provider>
   )
 }
